requirements.txt
Django==2.2
djongo==1.3.6
pymongo==3.12.1
sqlparse==0.2.4
requests==2.27.1
web3==5.29.0
py-solc-x==1.1.1